requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
Django==5.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.6.0
idna==3.10
jieba3k==0.35.1
joblib==1.4.2
lxml==5.3.2
lxml_html_clean==0.4.2
multidict==6.4.3
newspaper3k==0.2.8
nltk==3.9.1
pillow==11.2.1
propcache==0.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tinysegmenter==0.3
tldextract==5.2.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
